// job/src/shared/types/enums.ts
export enum URL {
    JobPost = "/api/v1/job/company/jobs"
 }

export enum TIME_FRAME {
    WEEKLY = 'week',
    MONTHLY = 'month',
    YEARLY = 'year',
} 

export enum ApplicationStatus {
    APPLIED = 'applied',
    IN_REVIEW = 'inreview',
    SHORTLISTED = 'shortlisted',
    INTERVIEW = 'interview',
    SELECTED = 'selected',
    REJECTED = 'rejected',
  }
  